+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- [Library] Elevation semenatic token `elevationBlurOverlayDefault`
+
+- [Library] Elevation semenatic tokens `elevationBlurOverlayDefault`, `elevationXOverlayDefault`
 
 ## [0.7.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.6.0...0.7.0) - 2024-11-29